draw curved edges

+++ Core/Util.cpp
@@ -90,7 +90,7 @@ double signedArea( const std::vector<XYZ>& v )
       return 0;
    double area2 = v[0].x * v.back().y - v.back().x * v[0].y;
 
-   for ( int i = 0; i+1 < v.size(); i++ )
+   for ( int i = 0; i+1 < (int)v.size(); i++ )
       area2 += v[i+1].x * v[i].y - v[i].x * v[i+1].y;
 
    return area2 / 2;
